+++ R/todo.R
@@ -26,6 +26,7 @@ todo_create<-function(rcm,subs,who="."){
   rcm<-rcm[grepl(who,rcm$hq_focal_point),]
   rcm<-rcm[!is.na(rcm$file.id),]
   rcm$submitter_emergency[is.na(rcm$submitter_emergency)]<-FALSE
+  rcm<-rcm[rcm$unit=="data",]
   todo <- rcm %>% todo_sort_by_priority
   class(todo)<-c("todo",class(todo))
   message(paste0("unrecognised items (new file id): ", nrow(subs[subs$new.file.id,])))